+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/update/abstracts/BaseEditMessageUpdate.kt
@@ -1,3 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.update.abstracts
 
-interface BaseEditMessageUpdate : BaseMessageUpdate
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.CommonMessage
+
+interface BaseEditMessageUpdate : BaseMessageUpdate {
+    override val data: CommonMessage<*>
+}